:recycle: Refactor: 도메인 인터페이스가 위치하는 패키지명 변경
