+++ composeApp/src/androidMain/kotlin/com/dh/ondot/core/util/AlarmService.kt
@@ -16,7 +16,7 @@ import androidx.core.app.ServiceCompat
 import co.touchlab.kermit.Logger
 import com.dh.ondot.R
 import com.dh.ondot.core.di.provideSoundPlayer
-import com.dh.ondot.domain.di.SoundPlayer
+import com.dh.ondot.domain.service.SoundPlayer
 import com.dh.ondot.domain.model.enums.AlarmType
 import com.dh.ondot.presentation.MainActivity
 import kotlinx.coroutines.CoroutineScope
