+++ composeApp/src/androidMain/kotlin/com/dh/ondot/core/util/AndroidAlarmStorage.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import com.dh.ondot.data.dataStore
-import com.dh.ondot.domain.di.AlarmStorage
+import com.dh.ondot.domain.service.AlarmStorage
 import com.dh.ondot.domain.model.response.AlarmDetail
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
