+++ composeApp/src/commonMain/kotlin/com/dh/ondot/domain/service/AlarmScheduler.kt
@@ -1,4 +1,4 @@
-package com.dh.ondot.domain.di
+package com.dh.ondot.domain.service
 
 import com.dh.ondot.domain.model.enums.AlarmType
 import com.dh.ondot.domain.model.response.AlarmDetail
