+++ composeApp/src/iosMain/kotlin/com/dh/ondot/core/util/IosSoundPlayer.kt
@@ -1,6 +1,6 @@
 package com.dh.ondot.core.util
 
-import com.dh.ondot.domain.di.SoundPlayer
+import com.dh.ondot.domain.service.SoundPlayer
 import kotlinx.cinterop.ExperimentalForeignApi
 import platform.AVFAudio.AVAudioPlayer
 import platform.AVFAudio.AVAudioSession
